Implement map.pretty library and use that for object details
This will do for now, but I think the way map objects are handled needs
a serious rethink; probably what I ultimately need is a clear separation
between the on-disk data and the in-memory data, with the latter
providing useful accessor methods and properly computing references to
objects/textures and whatnot. Then this can be a :pretty method on each
map/tile/object type.